pyproject.toml
[tool.poetry]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
flask = "*"
gunicorn = "^21.2.0"
flask-cors = "*"
pyttsx3 = "^2.98"
speechrecognition = "^3.11.0"
openai = "0.28.0"
waitress = "^3.0.2"
django-tts = "^0.3.1"
soundfile = "^0.12.1"
sounddevice = "^0.5.1"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"